Se finalizo la codificacion del login y CRUD del administrador

+++ FrontEnd/src/app/Componentes/mostrar/mostrar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {Router} from '@angular/router'; //importamos para las rutas
 import { Admin, ServAdminService } from './../../Services/serv-admin.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-mostrar',
@@ -8,6 +9,7 @@ import { Admin, ServAdminService } from './../../Services/serv-admin.service';
   styleUrls: ['./mostrar.component.css']
 })
 export class MostrarComponent implements OnInit {
+  
 
   ListarAdmin:Admin[]=[];
   constructor(private ServAdminService:ServAdminService, private router:Router) { }
@@ -25,18 +27,37 @@ export class MostrarComponent implements OnInit {
     );
   }
   eliminar(id:string){
-    this.ServAdminService.deleteAdmin(id).subscribe(
-      res=>{
-        console.log("admin eliminado")
-        this.listarAdmins()
-      },
-      err=>{
-        console.log(err)
+    Swal.fire({
+      title: 'Esta seguro que desea eliminar al Administrador?',
+      showDenyButton: true,
+      icon: 'warning',
+      //showCancelButton: true,
+      confirmButtonText: `Eliminar`,
+      denyButtonText: `Cancelar`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        this.ServAdminService.deleteAdmin(id).subscribe(
+          res=>{
+            console.log("Admin eliminado")
+            this.listarAdmins()
+          },
+          err=>{
+            console.log(err)
+          }
+        )
+        Swal.fire('Administrador eliminado', '', 'success')
       }
-    )
+    })
+    
   }
   modificar(id:string){
     this.router.navigate(['/edit/'+id])
+  }
+
+  outLogin(){
+    localStorage.clear()
+
   }
 
 }
